+++ packages/rocketchat-utils/server/index.js
@@ -7,6 +7,8 @@ import { roomTypes } from './lib/roomTypes';
 import { RoomTypeRouteConfig, RoomTypeConfig, RoomSettingsEnum, UiTextContext } from '../lib/RoomTypeConfig';
 import { RoomTypesCommon } from '../lib/RoomTypesCommon';
 import { isDocker } from './functions/isDocker';
+import { getAvatarUrlFromUsername } from '../lib/getAvatarUrlFromUsername';
+import { slashCommands } from '../lib/slashCommand';
 
 export {
 	t,
@@ -23,4 +25,6 @@ export {
 	RoomSettingsEnum,
 	UiTextContext,
 	isDocker,
+	getAvatarUrlFromUsername,
+	slashCommands,
 };
